+++ CHANGELOG.md
@@ -5,6 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- [lib] Directly adjusting image seek position no longer affects iteration with `ImageIterator` ([#42]).
+- [tui] Intensive performance drop while populating large image grids ([#41]).
+- [tui] Navigation across animated images ([#42]).
+  - No more waiting for the first frame to be rendered before moving on.
+
 ### Added
 - [lib] A common interface to be shared across all image classes ([#34]).
 - [lib] `BaseImage`, the baseclass of all image classes ([#34]).
@@ -14,11 +20,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [lib] `BaseImage.source_type` property ([#38]).
 - [lib] `KittyImage` class ([#39]).
 - [lib] Support for multiple render methods per render style via `BaseImage.set_render_method()` ([#39]).
+- [lib] Non-linear image iteration via `ImageIterator.seek()` ([#42]).
 - [cli] `kitty` render style choice for the `--style` CL option ([#39]).
+- [tui] Concurrent/Parallel frame rendering for TUI animations ([#42]).
 - [cli,tui] `--style` command-line option for render style selection ([#37]).
 - [lib,cli,tui] Support for the Kitty terminal graphics protocol ([#39]).
 - [lib,cli,tui] Automatic render style selection based on the detected terminal support ([#37]).
-
 
 ### Changed
 - [lib] `TermImage` is now a subclass of `BaseImage` ([#34]).
@@ -32,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#37]: https://github.com/AnonymouX47/term-image/pull/37
 [#38]: https://github.com/AnonymouX47/term-image/pull/38
 [#39]: https://github.com/AnonymouX47/term-image/pull/39
+[#41]: https://github.com/AnonymouX47/term-image/pull/41
+[#42]: https://github.com/AnonymouX47/term-image/pull/42
 
 
 ## [0.3.1] - 2022-05-04